requirements.txt
langchain
langchain-openai
openai
python-dotenv
langchain_community